#3625 - Remove DiamDebugEditor
- Removed DiamDebugEditor code
- Renamed package
