--- inception/inception-diam-editor/src/main/java/de/tudarmstadt/ukp/inception/experimental/editor/diamdebugeditor/DiamDebugEditor.html
@@ -1,25 +1,0 @@
-<!--
-  Licensed to the Technische Universität Darmstadt under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The Technische Universität Darmstadt 
-  licenses this file to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.
-   
-  http://www.apache.org/licenses/LICENSE-2.0
-  
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<html xmlns="http://www.w3.org/1999/xhtml"
-  xmlns:wicket="http://wicket.apache.org/dtds.data/wicket-xhtml1.4-strict.dtd">
-<body>
-  <wicket:panel>
-    <div wicket:id="vis" class="flex-content flex-v-container"/>
-  </wicket:panel>
-</body>
-</html>
