--- inception/inception-diam-editor/src/main/java/de/tudarmstadt/ukp/inception/experimental/editor/diamdebugeditor/DiamDebugEditorFactory.java
@@ -1,56 +1,0 @@
-/*
- * Licensed to the Technische Universität Darmstadt under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The Technische Universität Darmstadt 
- * licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.
- *  
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.tudarmstadt.ukp.inception.experimental.editor.diamdebugeditor;
-
-import org.apache.wicket.model.IModel;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Component;
-
-import de.tudarmstadt.ukp.clarin.webanno.api.CasProvider;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.paging.NoPagingStrategy;
-import de.tudarmstadt.ukp.inception.editor.AnnotationEditorBase;
-import de.tudarmstadt.ukp.inception.editor.AnnotationEditorFactoryImplBase;
-import de.tudarmstadt.ukp.inception.editor.action.AnnotationActionHandler;
-import de.tudarmstadt.ukp.inception.rendering.editorstate.AnnotatorState;
-
-@ConditionalOnExpression("${websocket.enabled:true}")
-@ConditionalOnProperty(prefix = "ui.diam-debugger", name = "enabled", havingValue = "true", matchIfMissing = false)
-@Component("diamDebugEditor")
-public class DiamDebugEditorFactory
-    extends AnnotationEditorFactoryImplBase
-{
-    @Override
-    public String getDisplayName()
-    {
-        return "DIAM Debugger";
-    }
-
-    @Override
-    public AnnotationEditorBase create(String aId, IModel<AnnotatorState> aModel,
-            AnnotationActionHandler aActionHandler, CasProvider aCasProvider)
-    {
-        return new DiamDebugEditor(aId, aModel, aActionHandler, aCasProvider);
-    }
-
-    @Override
-    public void initState(AnnotatorState aModelObject)
-    {
-        aModelObject.setPagingStrategy(new NoPagingStrategy());
-    }
-}
